refactor(web): split panel-manager into lifecycle and SSE parts
panel-manager.js sat at 449 of the 450 counted lines eslint's
max-lines allows, so any further change to the panel stack had to
begin by moving code out.

Move a panel's registration-to-enabled-entry path into
panel-lifecycle.js and the server-frame dispatch into panel-sse.js,
behind the injected-deps seam panel-placement.js already uses. The
core keeps every export and typedef and hands all four parts the same
state accessors, so no two modules can answer the same question
differently. Behavior unchanged.
